+++ tests/integration/cases/call_scalar_args_uniform_second_slot/V_heterogeneous_strategy_interface_call.v
@@ -1,3 +1,4 @@
+interface IVal {}
 interface ICallArg_ {}
 fn process(args ...ICallArg_) {}