+++ tests/integration/cases/call_snake_dotted_method/V_heterogeneous_strategy_interface_call.v
@@ -1,3 +1,4 @@
+interface IVal {}
 interface ICallArg_ {}
 struct Http_clientType_ {}
 fn (r Http_clientType_) fetch(args ...ICallArg_) {}