+++ tests/integration/cases/call_transform_no_wrapper/V_heterogeneous_strategy_interface_call.v
@@ -1,8 +1,9 @@
+interface IVal {}
 interface ICallArg_ {}
 fn process(args ...ICallArg_) {}
 
 fn main() {
-	process('hello', 'a');
-	process(42, 'b');
-	process(true, 'c');
+	process('hello');
+	process(42);
+	process(true);
 }