+++ tests/integration/cases/empty_orderedmap_with_sibling/V.v
@@ -1,9 +1,0 @@
-interface IVal {}
-
-fn main() {
-	my_data := [
-		IVal(map[string]IVal{}),
-		IVal([]IVal{}),
-	]
-	_ = my_data
-}