+++ tests/integration/cases/nested_list_with_empty_sibling/V_combined.v
@@ -2,14 +2,14 @@ interface IVal {}
 
 fn main() {
 	mut my_data := [
-		IVal([1, 2]),
-		IVal([]IVal{}),
-		IVal([3, 4]),
+		[1, 2],
+		[]int{},
+		[3, 4],
 	]
 	my_data = [
-		IVal([1, 2]),
-		IVal([]IVal{}),
-		IVal([3, 4]),
+		[1, 2],
+		[]int{},
+		[3, 4],
 	]
 	_ = my_data
 }