+++ tests/integration/cases/orderedmap_with_annotated_value/V.v
@@ -1,9 +1,0 @@
-interface IVal {}
-
-fn main() {
-	my_data := {
-		'a': IVal([]IVal{}),
-		'b': IVal(1),
-	}
-	_ = my_data
-}